Updating and fixing dependencies of the plugin
This change is part of [community #14019](https://tuleap.net/plugins/tracker/?aid=14019)

The goal of this commit is to recover a sane state where issuing a `mvn
hpi:run` command would successfully spin up a stable dev environment.

This includes:

  * Removing unnecessary dependency declaration for
    org.jenkins-ci.main.jenkins-core
  * Adding forgotten dependency on the Pipeline:Multibranch plugin
  * Updating targeted Jenkins version to the oldest supported LTS

⚠️  The [vulnerability warning](https://jenkins.io/security/advisory/2019-11-21/#SECURITY-1658)
is to be ignored as the Script Security Plugin is not one of our
dependencies, but one coming from Pipeline: Multibranch. One can
update the Script Security Plugin on its Jenkins instance, and this
shall be fixed in the next version (2.22) of the Multibranch plugin.
Once out, we'll update our dependency to it. ⚠️