Provide a way to add Tuleap Access Key to credentials
This change is part of [community #14019](https://tuleap.net/plugins/tracker/?aid=14019)

In order to test this change, on your Jenkins instance, you can go to
credentials management, add a new credential, and select Tuleap access
key. The provided key should be verified, if invalid or if it does not
have all the required scopes ("write:rest", "write:git_repository").