Rescan the repository if the branch given in the payload does not have its job
Part of [story #18320](https://tuleap.net/plugins/tracker/?aid=18320)
native support of pull requests in tuleap branch source jenkins plugin

If the job corresponding to the branch name of the payload then a scan
is launched at the repository level. There is only one scan and if the
branch is still not found, then a exception is raised in the log.

How to test :
 - Make sure to bind your git repository from  Tuleap with your Jenkins instance.
 - Push some content in a new branch
=> On the Jenkins side, a new scan of the repository is scheduled
=> After the repository scan, the new job/branch has been added and
builded ( at least it is still building ).

If you push in an already existing job/branch in Jenkins, there is no
scan of the repository.