Fix unappropriate error message left behind
This change fixes [request #21753](https://tuleap.net/plugins/tracker/?aid=21753)

The error message used to be there to prevent a faulty configuration of
the include/exclude which had not sane defaults. Now that the defaults
are sane again, we can remove it and introduce a proper project check on
the project selection box.