Add a branch discovery trait to TuleapSCMSource (#230)
* Add a branch discovery trait to TuleapSCMSource

This is part of [story #18320](https://tuleap.net/plugins/tracker/?aid=18320)

In the job configuration view of a Tuleap Project, you should have a
Branch discovery trait.

This trait should be present for any new job configuration by default.

Having this trait means that we will look for branches. If you remove
it, we don't anymore.