User should choose if they want to notify Tuleap.
[request #21858](https://tuleap.net/plugins/tracker/?tracker=140&aid=21858) Notifying Tuleap should not be systematic

How to test:
 - In your job configuration you now can add or remove the new behaviors: `Notify build status to Tuleap`
If the trait is selected then Jenkins will notify Tuleap.
If the traits is not selected then nothing will be done. Nothing will be
print in the Jenkins console of the build

This traits is not enabled by default, you have to select it if you want
notify Tuleap.