feat: Add the Tuleap Pull Request configuration trait
Part of [story#18320](https://tuleap.net/plugins/tracker/?aid=18320) native support of pull requests in tuleap branch source jenkins plugin

This is the introduction of the support of pull request in the plugin.
This contribution add a new configuration trait which allows the user to
retrieve the pull request (or not).

Note: this contribution add the traits only!

How to test:
 - Create or Update a Tuleap job by adding or removing the the `Tuleap
   Pull Requests autodiscovery` configuration trait
 - Launch the scan.
=> If the traits was added, then you should see a sentence in the Tuleap
Jenkins job log which says that the Pull request will be retrieved.
=> If the traits was not added or removed, then no sentence is displayed