Fixes [request #33320](https://tuleap.net/plugins/tracker/?aid=33320&group_id=101) Cannot setup a pipeline using a Tuleap Branch Source
Even if the `Tuleap` type can be selected in `Multibranch Pipeline` project type,
it never properly worked whatever the Jenkins version.
The implemented code did not take in account the fact that user can configure a `Tuleap` project in a `Multibranch Pipeline` project.
During the first implementation of the plugin, the creation of
`Mulitbranch Pipeline` projects was directly made by the "navigator" of
`Tuleap Organization Folder` project and other cases was not taken in
account.

Now, users should now be able to create a new Tuleap multibranch project.
No functional changes expected for Tuleap Organization project type

It should works fine with the Jenkins version 2.361.4 and later